refactor: extract logic from habits page into dedicated component
