+++ src/components/habit/index.ts
@@ -1,6 +1,3 @@
 export { default as AddHabitDialogButton } from './AddHabitDialogButton';
 export { default as EditHabitDialog } from './EditHabitDialog';
-export { default as HabitIcon } from './HabitIcon';
-export { default as HabitLastEntry } from './HabitLastEntry';
-export { default as HabitLongestStreak } from './HabitLongestStreak';
-export { default as HabitsTotalEntries } from './HabitTotalEntries';
+export { default as HabitsTable } from './HabitsTable';
