+++ src/hooks/use-file-field.ts
@@ -11,20 +11,21 @@ type ReturnValue = [
 const useFileField = (initialValue: FileFieldValue = null): ReturnValue => {
   const [value, setValue] = React.useState<FileFieldValue>(initialValue);
 
-  const handleChange = ({
-    target: { files },
-  }: React.ChangeEvent<HTMLInputElement>) => {
-    if (!files) {
-      return null;
-    }
+  const handleChange = React.useCallback(
+    ({ target: { files } }: React.ChangeEvent<HTMLInputElement>) => {
+      if (!files) {
+        return null;
+      }
 
-    const [value] = files;
-    setValue(value || null);
-  };
+      const [value] = files;
+      setValue(value || null);
+    },
+    []
+  );
 
-  const clearValue = () => {
+  const clearValue = React.useCallback(() => {
     setValue(null);
-  };
+  }, []);
 
   return [value, handleChange, clearValue];
 };
